+++ crates/biome_js_analyze/tests/specs/suspicious/noRedeclare/valid-index-signatures.ts
@@ -6,3 +6,15 @@ type ValidIndexSignatures = {
 		[index: string]: string;
 	};
 };
+
+interface I {
+	[index: number]: string;
+	[index: string]: string;
+}
+
+// See https://github.com/biomejs/biome/issues/175
+const x: { [key: string]: string } = {};
+let key;
+
+let a: { [key: string]: string };
+let b: { [key: string]: string };