+++ crates/biome_js_semantic/src/events.rs
@@ -326,7 +326,9 @@ impl SemanticEventExtractor {
             | TS_DECLARE_FUNCTION_DECLARATION
             | TS_DECLARE_FUNCTION_EXPORT_DEFAULT_DECLARATION
             | TS_METHOD_SIGNATURE_CLASS_MEMBER
-            | TS_METHOD_SIGNATURE_TYPE_MEMBER => {
+            | TS_METHOD_SIGNATURE_TYPE_MEMBER
+            | TS_INDEX_SIGNATURE_CLASS_MEMBER
+            | TS_INDEX_SIGNATURE_TYPE_MEMBER => {
                 self.push_scope(
                     node.text_range(),
                     ScopeHoisting::DontHoistDeclarationsToParent,
@@ -606,6 +608,8 @@ impl SemanticEventExtractor {
             | TS_DECLARE_FUNCTION_EXPORT_DEFAULT_DECLARATION
             | TS_METHOD_SIGNATURE_CLASS_MEMBER
             | TS_METHOD_SIGNATURE_TYPE_MEMBER
+            | TS_INDEX_SIGNATURE_CLASS_MEMBER
+            | TS_INDEX_SIGNATURE_TYPE_MEMBER
             | TS_INTERFACE_DECLARATION
             | TS_ENUM_DECLARATION
             | TS_TYPE_ALIAS_DECLARATION